# swagger.yaml
openapi: 3.0.3
info:
  title: Servana Backend API
  description: |
    API documentation for Servana - A customer service and chat management platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    Mobile endpoints use a separate authentication flow with OTP verification.
    
    ## Base URL
    - Development: `http://localhost:5000`
    - Production: Configure based on deployment
  version: 1.0.0
  contact:
    name: Servana API Support
    url: https://github.com/kurtbuset/backend_servana
servers:
  - url: http://localhost:5000
    description: Development server
  - url: http://192.168.137.77:5000
    description: Local network (Wi-Fi)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for web application authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
    Success:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
...
    Profile:
      type: object
      properties:
        prof_id:
          type: integer
        prof_firstname:
          type: string
        prof_middlename:
          type: string
        prof_lastname:
          type: string
        prof_address:
          type: string
        prof_date_of_birth:
          type: string
          format: date
        prof_created_at:
          type: string
          format: date-time
    Department:
      type: object
      properties:
        dept_id:
          type: integer
        dept_name:
          type: string
        dept_is_active:
          type: boolean
        dept_created_at:
          type: string
          format: date-time
        dept_created_by:
          type: integer
...
paths:
...
  /profile:
    get:
      tags:
        - Profile
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sys_user_id:
                        type: integer
                      sys_user_email:
                        type: string
                      role_id:
                        type: integer
                      role_name:
                        type: string
                      privilege:
                        type: object
                      departments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Department'
                      profile:
                        $ref: '#/components/schemas/Profile'
                      image:
                        type: object

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Profile
      summary: Update current user profile
      description: Update the authenticated user's profile information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                middleName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                address:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profile/image:
    post:
      tags:
        - Profile
      summary: Upload profile image
      description: Upload a profile picture for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary

      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      img_location:
                        type: string
                      image:
                        type: object
        '400':
          description: No image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profile/agent-status:
    get:
      tags:
        - Profile
      summary: Get agent status
      description: Get the current agent's online/offline status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agent status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      agent_status:
                        type: string
                        enum: [online, offline, away]
    
    put:
      tags:
        - Profile
      summary: Update agent status
      description: Update the agent's online/offline status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_status
              properties:
                agent_status:
                  type: string
                  enum: [online, offline, away]
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      agent_status:
                        type: string
...
  /chat/messages/{chat_group_id}:
    get:
      tags:
        - Chat
      summary: Get chat messages
      description: Retrieve all messages for a specific chat group
      security:
        - bearerAuth: []
      parameters:
        - name: chat_group_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
...
  /messages:
    get:
      tags:
        - Mobile - Messages
      summary: Get mobile messages
      description: Retrieve messages for mobile client
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
...